+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New option `--lang` to `odood translate regenerate` command.
   With this option, translation file will be detected automatically.
+- New option `--repeat` to `odood db populate` command, that allow to repeat database population N times.
 
 ---